requirements.txt
tushare>=1.2.89
pandas>=1.3.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
pyyaml>=6.0
